refactor: 글 메타데이터 정보 가져오는 방식
