+++ apps/blog/app/(n)/[category]/[sub]/[slug]/page.tsx
@@ -1,9 +1,10 @@
+import { getSlugMetadata } from "@/constants/notes";
 import { SlugParams } from "@/constants/params.types";
 
 export async function generateMetadata({ params }: { params: Promise<SlugParams> }) {
   const { category, sub, slug } = await params;
 
-  const { metadata } = await import(`@/content/${category}/${sub}/${slug}/page.mdx`);
+  const metadata = await getSlugMetadata(category, sub, slug);
 
   return { ...metadata, authors: [{ name: "Kihyun Ryu" }], creator: "Kihyun Ryu", publisher: "Kihyun Ryu" };
 }
