--- apps/blog/constants/metadata.types.ts
@@ -1,9 +1,0 @@
-import type { Metadata } from "next";
-
-export type NoteMetadata = Partial<Metadata> & {
-  other: {
-    status: "ready" | "draft";
-    createdAt: string;
-    updatedAt?: string;
-  };
-};
