+++ apps/blog/constants/notes.ts
@@ -1,4 +1,4 @@
-import { Category, Notes } from "@/constants/notes.types";
+import { Category, NoteMetadata, Notes } from "@/constants/notes.types";
 import { getFolderNames } from "@/lib/file-system";
 import { isCategory } from "@/lib/type-guards";
 import path from "path";
@@ -13,6 +13,13 @@ export const getSubCategoriesByCategory = async (category: Category) => await ge
 
 /** 글 제목 */
 export const getSlugsByCategoryAndSub = async (category: Category, sub: string) => await getFolderNames(path.join(CONTENT_PATH, category, sub));
+
+// TODO: metadata의 타입가드 및 불일치시 에러 던지기
+/** 글 정보 */
+export const getSlugMetadata = async (category: Category, sub: string, slug: string): Promise<NoteMetadata> => {
+  const { metadata } = await import(`@/content/${category}/${sub}/${slug}/page.mdx`);
+  return metadata;
+};
 
 /** 노트 */
 export const NOTES: Notes = await (async () => {
