+++ apps/blog/constants/notes.types.ts
@@ -1,7 +1,17 @@
+import { Metadata } from "next";
+
 export type Category = string;
 
 export type Notes = {
   [category: string]: {
     [sub: string]: string[];
+  };
+};
+
+export type NoteMetadata = Partial<Metadata> & {
+  other: {
+    status: "ready" | "draft";
+    createdAt: string;
+    updatedAt?: string;
   };
 };
